Refactoring and adding more comments
