--- core/src/main/java/org/polypheny/db/document/rules/DocumentValues.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2019-2021 The Polypheny Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.polypheny.db.document.rules;
-
-import com.google.common.collect.ImmutableList;
-import org.polypheny.db.plan.RelOptCluster;
-import org.polypheny.db.plan.RelTraitSet;
-import org.polypheny.db.rel.RelInput;
-import org.polypheny.db.rel.logical.LogicalValues;
-import org.polypheny.db.rel.type.RelDataType;
-import org.polypheny.db.rex.RexLiteral;
-
-public class DocumentValues extends LogicalValues {
-
-    public DocumentValues( RelOptCluster cluster, RelTraitSet traitSet, RelDataType rowType, ImmutableList<ImmutableList<RexLiteral>> tuples ) {
-        super( cluster, traitSet, rowType, tuples );
-    }
-
-
-    public DocumentValues( RelInput input ) {
-        super( input );
-    }
-
-}
