+++ dbms/src/main/java/org/polypheny/db/processing/AbstractQueryProcessor.java
@@ -60,7 +60,7 @@ import org.polypheny.db.catalog.exceptions.UnknownDatabaseException;
 import org.polypheny.db.catalog.exceptions.UnknownSchemaException;
 import org.polypheny.db.catalog.exceptions.UnknownTableException;
 import org.polypheny.db.config.RuntimeConfig;
-import org.polypheny.db.document.DataModelShuttle;
+import org.polypheny.db.document.util.DataModelShuttle;
 import org.polypheny.db.information.InformationGroup;
 import org.polypheny.db.information.InformationManager;
 import org.polypheny.db.information.InformationPage;
